requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dlib==19.24.8
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
requests==2.32.3
setuptools==80.0.0
tinify==1.7.0
urllib3==2.4.0
Werkzeug==3.1.3
